feat: Enhance transaction handling and user management
- Improved transaction creation logic in TransactionService to handle insufficient funds and create transaction records on failure.
- Updated FetchUser function in UserService to use user ID directly instead of uint32.
- Modified WebhookService to standardize notification types for transactions.
- Enhanced GetUserTotalBalance to consider user settings for default currency.
- Refactored error handling in utils/helpers.go for better clarity and structure.
- Introduced new models for LoginActivity, PlatformSetting, SavedRecipient, and Withdrawal to support additional features.
- Added admin routes for user and transaction management in the admin API.
- Created new types for wallet settings to streamline request handling.

+++ types/User.go
@@ -27,7 +27,7 @@ type UserWallet struct {
 }
 
 type UserResponse struct {
-	ID             int64
+	ID             uint
 	UserID         uint32             `json:"user_id"`
 	Email          string             `json:"email"`
 	ProfilePicture string             `json:"profile_picture"`
